+++ CHANGELOG.md
@@ -9,6 +9,12 @@ All releases must adhere to [SemVer 2](https://semver.org/) naming convention an
 >PATCH version - when you make backwards compatible bug fixes.
 
 ## LivingMarkup [Unreleased]
+Released: TBD. Notable changes:
+* Added: ProcessorFactory.
+* Added: addRoutine().
+* Removed: addMethod().
+* Added: setups and tear downs for PHPUnit tests.
+* Added: Config version 3 support with Element and Routines.
 
 ## LivingMarkup 2.0.0
 Released: 2021-09-01. Notable changes: